third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-video-element/video-loading-load-deferred.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Video with loading=lazy that is not visible in viewport does not load video data
  assert_equals: expected 0 but got 4
Harness: the test ran to completion.
